SISRP-29960 - Rename Current Enrollments API
